Conflicts that have been resolved:
	insts/bgg/PFSCHED/PFSCHED.cpp
	insts/joel/jfuncs/jfuncs.c
	src/audio/audio_devices.cpp
	src/rtcmix/InputFile.cpp
	src/rtcmix/Makefile
	src/rtcmix/RTOption.cpp
	src/rtcmix/RTcmix.cpp
	src/rtcmix/RTcmixMain.cpp
	src/rtcmix/RTcmix_API.h
	src/rtcmix/TaskManager.cpp
	src/rtcmix/addcheckfunc.cpp
	src/rtcmix/atomic_stack.h
	src/rtcmix/command_line.c
	src/rtcmix/main.cpp
	src/rtcmix/maxmspstuff/INSTDOBJS
	src/rtcmix/profile.c
	src/rtcmix/rtinput.cpp
	src/rtcmix/rtprofile.cpp

+++ insts/jg/denoise/denoise.c
@@ -85,7 +85,7 @@ static void malerr(char *, int);
 
 
 double
-denoise(float p[], int n_args)
+denoise(double p[], int n_args)
 {
    float *input,               /* input buffer */
    *output,                    /* output buffer */
